+++ migration/src/m20260524_001051_action.rs
@@ -26,9 +26,11 @@ impl MigrationTrait for Migration {
     }
 
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .drop_table(Table::drop().table(Action::Table).if_exists().to_owned())
-            .await?;
+        if std::env::var("ENVIRONMENT").unwrap() != "production" {
+            manager
+                .drop_table(Table::drop().table(Action::Table).if_exists().to_owned())
+                .await?;
+        }
         Ok(())
     }
 }
