+++ migration/src/m20260524_001135_box.rs
@@ -94,17 +94,19 @@ impl MigrationTrait for Migration {
     }
 
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
-        manager
-            .drop_table(
-                Table::drop()
-                    .table(Boxhistory::Table)
-                    .if_exists()
-                    .to_owned(),
-            )
-            .await?;
-        manager
-            .drop_table(Table::drop().table(Rvbox::Table).if_exists().to_owned())
-            .await?;
+        if std::env::var("ENVIRONMENT").unwrap() != "production" {
+            manager
+                .drop_table(
+                    Table::drop()
+                        .table(Boxhistory::Table)
+                        .if_exists()
+                        .to_owned(),
+                )
+                .await?;
+            manager
+                .drop_table(Table::drop().table(Rvbox::Table).if_exists().to_owned())
+                .await?;
+        }
         Ok(())
     }
 }
